refactor: split memory_store into helpers + tests subfiles
