+++ src/memory_store.rs
@@ -661,965 +661,8 @@ impl MemoryRecordFileStore {
     }
 }
 
-fn default_record_path_for_vdb_uri(uri: &str) -> PathBuf {
-    let db_path = PathBuf::from(uri);
-    if db_path.file_name().and_then(|value| value.to_str()) == Some("lancedb") {
-        return db_path
-            .parent()
-            .unwrap_or_else(|| Path::new("."))
-            .join("memories")
-            .join("records.json");
-    }
-    db_path.join("memory_records.json")
-}
-
-fn upsert_record_sync(
-    path: PathBuf,
-    lock_path: PathBuf,
-    cache: Arc<Mutex<MemoryRecordCache>>,
-    record: MemoryRecord,
-) -> Result<()> {
-    let _lock = AdvisoryFileLock::acquire(lock_path)?;
-    let mut file = PersistedMemoryRecordFile {
-        records: load_records_sync(&path)?,
-        ..Default::default()
-    };
-    if let Some(existing) = file.records.iter_mut().find(|item| item.id == record.id) {
-        *existing = record;
-    } else {
-        file.records.push(record);
-    }
-    write_record_file_sync(&path, &file)?;
-    let records = record_map_from_records(file.records);
-    let state = record_file_state(&path)?;
-    update_record_cache(&cache, state, records);
-    Ok(())
-}
-
-fn update_record_sync<F>(
-    path: PathBuf,
-    lock_path: PathBuf,
-    cache: Arc<Mutex<MemoryRecordCache>>,
-    id: &str,
-    update: F,
-) -> Result<MemoryRecord>
-where
-    F: FnOnce(&mut MemoryRecord) -> bool,
-{
-    let _lock = AdvisoryFileLock::acquire(lock_path)?;
-    let mut file = PersistedMemoryRecordFile {
-        records: load_records_sync(&path)?,
-        ..Default::default()
-    };
-    let Some(record) = file.records.iter_mut().find(|item| item.id == id) else {
-        bail!("memory record not found: {id}");
-    };
-    let changed = update(record);
-    let updated = record.clone();
-    if changed {
-        write_record_file_sync(&path, &file)?;
-        let records = record_map_from_records(file.records);
-        let state = record_file_state(&path)?;
-        update_record_cache(&cache, state, records);
-    }
-    Ok(updated)
-}
-
-fn delete_record_sync(
-    path: PathBuf,
-    lock_path: PathBuf,
-    cache: Arc<Mutex<MemoryRecordCache>>,
-    id: &str,
-) -> Result<Option<MemoryRecord>> {
-    let _lock = AdvisoryFileLock::acquire(lock_path)?;
-    let mut file = PersistedMemoryRecordFile {
-        records: load_records_sync(&path)?,
-        ..Default::default()
-    };
-    let Some(index) = file.records.iter().position(|record| record.id == id) else {
-        return Ok(None);
-    };
-    let deleted = file.records.remove(index);
-    write_record_file_sync(&path, &file)?;
-    let records = record_map_from_records(file.records);
-    let state = record_file_state(&path)?;
-    update_record_cache(&cache, state, records);
-    Ok(Some(deleted))
-}
-
-fn load_record_map_cached_sync(
-    path: &Path,
-    cache: &Arc<Mutex<MemoryRecordCache>>,
-) -> Result<HashMap<String, MemoryRecord>> {
-    let state = record_file_state(path)?;
-    {
-        let cache = cache.lock().expect("memory record cache lock poisoned");
-        if cache.state == Some(state) {
-            return Ok(cache.records.clone());
-        }
-    }
-
-    let records = record_map_from_records(load_records_sync(path)?);
-    update_record_cache(cache, state, records.clone());
-    Ok(records)
-}
-
-fn get_record_cached_sync(
-    path: &Path,
-    cache: &Arc<Mutex<MemoryRecordCache>>,
-    id: &str,
-) -> Result<Option<MemoryRecord>> {
-    Ok(load_record_map_cached_sync(path, cache)?.get(id).cloned())
-}
-
-fn update_record_cache(
-    cache: &Arc<Mutex<MemoryRecordCache>>,
-    state: MemoryRecordFileState,
-    records: HashMap<String, MemoryRecord>,
-) {
-    let mut cache = cache.lock().expect("memory record cache lock poisoned");
-    cache.state = Some(state);
-    cache.records = records;
-}
-
-fn record_map_from_records(records: Vec<MemoryRecord>) -> HashMap<String, MemoryRecord> {
-    records
-        .into_iter()
-        .map(|record| (record.id.clone(), record))
-        .collect()
-}
-
-fn record_file_state(path: &Path) -> Result<MemoryRecordFileState> {
-    match fs::metadata(path) {
-        Ok(metadata) => Ok(MemoryRecordFileState::Present {
-            modified_at: metadata.modified().ok(),
-            len: metadata.len(),
-        }),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
-            Ok(MemoryRecordFileState::Missing)
-        }
-        Err(err) => Err(err).with_context(|| format!("stat memory records {}", path.display())),
-    }
-}
-
-fn load_records_sync(path: &Path) -> Result<Vec<MemoryRecord>> {
-    let file = match File::open(path) {
-        Ok(file) => file,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(Vec::new()),
-        Err(err) => {
-            return Err(err).with_context(|| format!("read memory records {}", path.display()));
-        }
-    };
-    let reader = BufReader::new(file);
-    match serde_json::from_reader::<_, PersistedMemoryRecordEnvelope>(reader) {
-        Ok(PersistedMemoryRecordEnvelope::Versioned(file)) => Ok(file.records),
-        Ok(PersistedMemoryRecordEnvelope::Legacy(records)) => Ok(records),
-        Err(err) if err.is_eof() => Ok(Vec::new()),
-        Err(err) => Err(err).with_context(|| format!("parse memory records {}", path.display())),
-    }
-}
-
-fn write_record_file_sync(path: &Path, file: &PersistedMemoryRecordFile) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("create memory records dir {}", parent.display()))?;
-    }
-    let tmp_path = path.with_extension(format!("json.tmp-{}", uuid::Uuid::new_v4()));
-    let tmp_file = File::create(&tmp_path)
-        .with_context(|| format!("create memory records temp file {}", tmp_path.display()))?;
-    let mut writer = BufWriter::new(tmp_file);
-    serde_json::to_writer(&mut writer, file)
-        .with_context(|| format!("serialize memory records {}", tmp_path.display()))?;
-    writer
-        .flush()
-        .with_context(|| format!("flush memory records temp file {}", tmp_path.display()))?;
-    if let Err(err) = atomic_file::replace_file(&tmp_path, path) {
-        let _ = fs::remove_file(&tmp_path);
-        return Err(err).with_context(|| format!("replace memory records {}", path.display()));
-    }
-    Ok(())
-}
-
-fn memory_record_for_hit(
-    records: &HashMap<String, MemoryRecord>,
-    metadata: &MemoryMetadata,
-) -> Option<MemoryRecord> {
-    match metadata.id.as_ref() {
-        Some(id) => records.get(id).cloned(),
-        None => Some(MemoryRecord::from(metadata.clone())),
-    }
-}
-
-fn normalize_memory_record_patch(mut patch: MemoryRecordPatch) -> Result<MemoryRecordPatch> {
-    if let Some(title) = patch.title.take() {
-        patch.title = Some(title.trim().to_string());
-    }
-    if let Some(content) = patch.content.take() {
-        let content = content.trim().to_string();
-        if content.is_empty() {
-            bail!("memory content must not be empty");
-        }
-        patch.content = Some(content);
-    }
-    if let Some(labels) = patch.labels.take() {
-        patch.labels = Some(normalized_labels(labels));
-    }
-    if let Some(importance) = patch.importance.take() {
-        patch.importance = Some(clamp_importance(importance));
-    }
-    if let Some(session_id) = patch.session_id.take() {
-        patch.session_id = Some(normalized_optional_id(session_id));
-    }
-    if let Some(thread_id) = patch.thread_id.take() {
-        patch.thread_id = Some(normalized_optional_id(thread_id));
-    }
-    Ok(patch)
-}
-
-fn patch_requires_index_refresh(patch: &MemoryRecordPatch) -> bool {
-    patch.content.is_some()
-        || patch.scope.is_some()
-        || patch.session_id.is_some()
-        || patch.thread_id.is_some()
-}
-
-fn apply_memory_record_patch(record: &mut MemoryRecord, patch: MemoryRecordPatch) -> bool {
-    let before = record.clone();
-    if let Some(title) = patch.title.filter(|title| !title.is_empty()) {
-        record.title = title;
-    }
-    if let Some(content) = patch.content {
-        record.content = content;
-    }
-    if let Some(labels) = patch.labels {
-        record.labels = labels;
-    }
-    if let Some(importance) = patch.importance {
-        record.importance = importance;
-    }
-    if let Some(pinned) = patch.pinned {
-        record.pinned = pinned;
-    }
-    if let Some(scope) = patch.scope {
-        record.scope = scope;
-    }
-    if let Some(session_id) = patch.session_id {
-        record.session_id = session_id;
-    }
-    if let Some(thread_id) = patch.thread_id {
-        record.thread_id = thread_id;
-    }
-    if let Some(source_span) = patch.source_span {
-        record.source_span = source_span;
-    }
-    if *record != before {
-        record.updated_at_unix_seconds = unix_timestamp_seconds();
-        true
-    } else {
-        false
-    }
-}
-
-fn list_label_counts<'a>(
-    records: impl IntoIterator<Item = &'a MemoryRecord>,
-    scope: Option<&MemoryScope>,
-) -> Vec<MemoryLabelCount> {
-    let mut counts = HashMap::<MemoryLabel, usize>::new();
-    for record in records {
-        if scope.is_some_and(|scope| &record.scope != scope) {
-            continue;
-        }
-        for label in &record.labels {
-            *counts.entry(label.clone()).or_default() += 1;
-        }
-    }
-    let mut counts = counts
-        .into_iter()
-        .map(|(label, count)| MemoryLabelCount { label, count })
-        .collect::<Vec<_>>();
-    counts.sort_by(|left, right| {
-        right.count.cmp(&left.count).then_with(|| {
-            memory_label_sort_key(&left.label).cmp(memory_label_sort_key(&right.label))
-        })
-    });
-    counts
-}
-
-fn memory_label_sort_key(label: &MemoryLabel) -> &'static str {
-    match label {
-        MemoryLabel::Insight => "insight",
-        MemoryLabel::Decision => "decision",
-        MemoryLabel::Fact => "fact",
-        MemoryLabel::Procedure => "procedure",
-        MemoryLabel::Experience => "experience",
-    }
-}
-
-fn normalized_labels(labels: Vec<MemoryLabel>) -> Vec<MemoryLabel> {
-    if labels.is_empty() {
-        return vec![MemoryLabel::Experience];
-    }
-    labels
-}
-
-fn normalized_optional_id(value: Option<String>) -> Option<String> {
-    value
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-}
-
-fn required_memory_id(value: String, field: &str) -> Result<String> {
-    let value = value.trim().to_string();
-    if value.is_empty() {
-        bail!("{field} is required for scoped memory promotion");
-    }
-    Ok(value)
-}
-
-fn clamp_importance(importance: f32) -> f32 {
-    if importance.is_nan() {
-        return DEFAULT_IMPORTANCE;
-    }
-    importance.clamp(0.1, 1.0)
-}
-
-fn memory_scope_key(scope: &MemoryScope) -> &'static str {
-    match scope {
-        MemoryScope::User => "user",
-        MemoryScope::Workspace => "workspace",
-        MemoryScope::Project => "project",
-        MemoryScope::Thread => "thread",
-        MemoryScope::Session => "session",
-    }
-}
-
-fn memory_scope_from_key(value: &str) -> MemoryScope {
-    match value {
-        "user" => MemoryScope::User,
-        "workspace" => MemoryScope::Workspace,
-        "thread" => MemoryScope::Thread,
-        "session" => MemoryScope::Session,
-        _ => MemoryScope::Project,
-    }
-}
-
-fn title_from_content(content: &str) -> String {
-    let first_line = content
-        .lines()
-        .find(|line| !line.trim().is_empty())
-        .unwrap_or("Memory");
-    let title = first_line
-        .split_terminator(['.', '!', '?'])
-        .next()
-        .unwrap_or(first_line)
-        .trim();
-    truncate_chars(title, 80)
-}
-
-fn truncate_chars(value: &str, max_chars: usize) -> String {
-    let mut chars = value.chars();
-    let truncated = chars.by_ref().take(max_chars).collect::<String>();
-    if chars.next().is_some() {
-        format!("{truncated}...")
-    } else {
-        truncated
-    }
-}
-
-fn unix_timestamp_seconds() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs()
-}
-
+include!("memory_store_helpers.rs");
 #[cfg(test)]
 mod tests {
-    use anyhow::{Result, anyhow};
-    use async_trait::async_trait;
-    use serde_json::Value;
-
-    use super::*;
-    use crate::agent::Message;
-    use crate::llm::{
-        ContentBlock, EmbeddingBackend, EmbeddingInputKind, LlmBackend, LlmResponse, MockLlm,
-    };
-
-    struct FailingLlmBackend;
-
-    #[async_trait]
-    impl LlmBackend for FailingLlmBackend {
-        async fn ask(&self, _messages: &[Message], _tools: &[Value]) -> Result<LlmResponse> {
-            Ok(LlmResponse {
-                content: vec![ContentBlock::Text {
-                    text: "ok".to_string(),
-                }],
-                stop_reason: Some("end_turn".to_string()),
-                usage: None,
-            })
-        }
-
-        async fn embed(&self, _text: &str) -> Result<Vec<f32>> {
-            Err(anyhow!("llm embedding path should stay unused"))
-        }
-
-        async fn summarize(&self, _messages: &[Message], _instruction: &str) -> Result<String> {
-            Ok("summary".to_string())
-        }
-    }
-
-    struct FixedEmbeddingBackend;
-
-    #[async_trait]
-    impl EmbeddingBackend for FixedEmbeddingBackend {
-        async fn embed(&self, text: &str, kind: EmbeddingInputKind) -> Result<Vec<f32>> {
-            let vector = match kind {
-                EmbeddingInputKind::Query if text.contains("parser") => vec![1.0, 0.0, 0.0, 0.0],
-                EmbeddingInputKind::Document if text.contains("DeepSeek") => {
-                    vec![1.0, 0.0, 0.0, 0.0]
-                }
-                _ => vec![0.0, 1.0, 0.0, 0.0],
-            };
-            Ok(vector)
-        }
-    }
-
-    fn test_memory_record(id: &str, content: &str) -> MemoryRecord {
-        MemoryRecord {
-            id: id.to_string(),
-            title: title_from_content(content),
-            content: content.to_string(),
-            labels: vec![MemoryLabel::Fact],
-            importance: DEFAULT_IMPORTANCE,
-            pinned: false,
-            source: MemorySource::UserCreated,
-            scope: MemoryScope::Project,
-            session_id: None,
-            thread_id: None,
-            source_span: None,
-            created_at_unix_seconds: 1,
-            updated_at_unix_seconds: 1,
-        }
-    }
-
-    #[test]
-    fn memory_record_file_store_writes_compact_json_and_reads_legacy_records() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let path = temp.path().join("records.json");
-        let legacy_path = temp.path().join("legacy_records.json");
-        let record = test_memory_record("memory-test-1", "Compact storage should stay readable.");
-        let file = PersistedMemoryRecordFile {
-            version: MEMORY_RECORDS_FILE_VERSION,
-            records: vec![record.clone()],
-        };
-
-        write_record_file_sync(&path, &file).expect("write compact record file");
-        let content = fs::read_to_string(&path).expect("read compact record file");
-        assert!(!content.contains('\n'));
-        assert_eq!(
-            load_records_sync(&path).expect("load compact records"),
-            vec![record.clone()]
-        );
-
-        fs::write(
-            &legacy_path,
-            serde_json::to_string(&vec![record.clone()]).expect("serialize legacy records"),
-        )
-        .expect("write legacy record file");
-        assert_eq!(
-            load_records_sync(&legacy_path).expect("load legacy records"),
-            vec![record]
-        );
-    }
-
-    #[test]
-    fn memory_record_index_scope_key_respects_memory_scope() {
-        let mut workspace = test_memory_record(
-            "memory-workspace",
-            "Workspace-scoped records may carry thread provenance.",
-        );
-        workspace.scope = MemoryScope::Workspace;
-        workspace.session_id = Some("session-1".to_string());
-        workspace.thread_id = Some("thread-1".to_string());
-        assert_eq!(workspace.index_scope_key(), "workspace");
-
-        let mut thread = workspace.clone();
-        thread.scope = MemoryScope::Thread;
-        assert_eq!(thread.index_scope_key(), "thread-1");
-
-        let mut session = workspace.clone();
-        session.scope = MemoryScope::Session;
-        assert_eq!(session.index_scope_key(), "session-1");
-
-        let mut project = workspace;
-        project.scope = MemoryScope::Project;
-        assert_eq!(project.index_scope_key(), "project");
-    }
-
-    #[test]
-    fn memory_promotion_base_enforces_scope_rules() {
-        let workspace = NewMemoryRecord::promotion_base(
-            MemorySource::ProtocolWrite,
-            MemoryPromotionTarget::Workspace {
-                session_id: Some(" session-1 ".to_string()),
-                thread_id: Some(" thread-1 ".to_string()),
-            },
-        )
-        .expect("workspace promotion");
-        assert_eq!(workspace.scope, MemoryScope::Workspace);
-        assert_eq!(workspace.session_id.as_deref(), Some("session-1"));
-        assert_eq!(workspace.thread_id.as_deref(), Some("thread-1"));
-        assert_eq!(workspace.source_span, None);
-
-        let thread_err = NewMemoryRecord::promotion_base(
-            MemorySource::ThreadDistill,
-            MemoryPromotionTarget::Thread {
-                session_id: Some("session-1".to_string()),
-                thread_id: " ".to_string(),
-                source_span: None,
-            },
-        )
-        .expect_err("thread promotion needs a thread id");
-        assert!(thread_err.to_string().contains("thread_id is required"));
-
-        let session = NewMemoryRecord::promotion_base(
-            MemorySource::SessionDistill,
-            MemoryPromotionTarget::Session {
-                session_id: " session-2 ".to_string(),
-                source_span: Some(MemorySourceSpan {
-                    start_turn_index: 2,
-                    end_turn_index: 4,
-                }),
-            },
-        )
-        .expect("session promotion");
-        assert_eq!(session.scope, MemoryScope::Session);
-        assert_eq!(session.session_id.as_deref(), Some("session-2"));
-        assert_eq!(session.thread_id.as_deref(), Some("session-2"));
-        assert_eq!(
-            session.source_span,
-            Some(MemorySourceSpan {
-                start_turn_index: 2,
-                end_turn_index: 4,
-            })
-        );
-    }
-
-    #[test]
-    fn memory_record_file_store_refreshes_cache_when_file_changes() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let path = temp.path().join("records.json");
-        let cache = Arc::new(Mutex::new(MemoryRecordCache::default()));
-        let first = test_memory_record("memory-cache-1", "First cached record.");
-        let second = test_memory_record(
-            "memory-cache-2",
-            "Second cached record with a longer payload.",
-        );
-
-        write_record_file_sync(
-            &path,
-            &PersistedMemoryRecordFile {
-                version: MEMORY_RECORDS_FILE_VERSION,
-                records: vec![first.clone()],
-            },
-        )
-        .expect("write first record file");
-        let first_map = load_record_map_cached_sync(&path, &cache).expect("load first cache");
-        assert_eq!(first_map.get(&first.id), Some(&first));
-
-        write_record_file_sync(
-            &path,
-            &PersistedMemoryRecordFile {
-                version: MEMORY_RECORDS_FILE_VERSION,
-                records: vec![second.clone()],
-            },
-        )
-        .expect("write second record file");
-        let second_map = load_record_map_cached_sync(&path, &cache).expect("refresh cache");
-        assert_eq!(second_map.get(&second.id), Some(&second));
-        assert!(!second_map.contains_key(&first.id));
-    }
-
-    #[tokio::test]
-    async fn memory_store_inserts_and_searches_memory_records() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let store = MemoryStore::new(
-            Arc::new(MockLlm),
-            Arc::new(VectorDB::new(temp.path().to_str().expect("utf8 path"))),
-        );
-
-        let saved = store
-            .insert(NewMemoryRecord::experience(
-                "DeepSeek DSML requires a structured parser.",
-            ))
-            .await
-            .expect("insert memory");
-        assert!(saved.id.starts_with("memory-"));
-        assert_eq!(saved.title, "DeepSeek DSML requires a structured parser");
-        assert_eq!(saved.labels, vec![MemoryLabel::Experience]);
-
-        let hits = store
-            .search("DSML parser", 8)
-            .await
-            .expect("search memories");
-        assert_eq!(hits.len(), 1);
-        assert_eq!(hits[0].record.id, saved.id);
-        assert_eq!(
-            hits[0].record.content,
-            "DeepSeek DSML requires a structured parser."
-        );
-        assert_eq!(
-            store.get(&saved.id).await.expect("get saved memory"),
-            Some(saved)
-        );
-    }
-
-    #[tokio::test]
-    async fn memory_store_uses_separate_embedding_backend_for_vector_paths() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let store = MemoryStore::new_with_embedding_backend(
-            Arc::new(FailingLlmBackend),
-            Arc::new(FixedEmbeddingBackend),
-            Arc::new(VectorDB::new(temp.path().to_str().expect("utf8 path"))),
-        );
-
-        let saved = store
-            .insert(NewMemoryRecord::experience(
-                "DeepSeek DSML requires a structured parser.",
-            ))
-            .await
-            .expect("insert memory with separate embedding backend");
-        let hits = store
-            .search("structured parser", 8)
-            .await
-            .expect("search memories with separate embedding backend");
-
-        assert_eq!(hits.len(), 1);
-        assert_eq!(hits[0].record.id, saved.id);
-    }
-
-    #[tokio::test]
-    async fn memory_store_clamps_importance_and_defaults_labels() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let store = MemoryStore::new(
-            Arc::new(MockLlm),
-            Arc::new(VectorDB::new(temp.path().to_str().expect("utf8 path"))),
-        );
-
-        let saved = store
-            .insert(NewMemoryRecord {
-                title: Some("".to_string()),
-                content: "A durable fact".to_string(),
-                labels: Vec::new(),
-                importance: 4.0,
-                pinned: false,
-                source: MemorySource::UserCreated,
-                scope: MemoryScope::Workspace,
-                session_id: None,
-                thread_id: None,
-                source_span: None,
-            })
-            .await
-            .expect("insert memory");
-
-        assert_eq!(saved.importance, 1.0);
-        assert_eq!(saved.labels, vec![MemoryLabel::Experience]);
-        assert_eq!(saved.scope, MemoryScope::Workspace);
-    }
-
-    #[tokio::test]
-    async fn memory_store_persists_thread_provenance_across_instances() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let db_path = temp.path().join("lancedb");
-        let record_path = temp.path().join("memories").join("records.json");
-        let backend = Arc::new(MockLlm);
-        let vdb = Arc::new(VectorDB::new(db_path.to_str().expect("utf8 path")));
-        let store =
-            MemoryStore::new_with_record_path(backend.clone(), vdb.clone(), record_path.clone());
-
-        let saved = store
-            .insert(NewMemoryRecord {
-                title: Some("Thread decision".to_string()),
-                content: "Keep memory retrieval behind MemoryStore.".to_string(),
-                labels: vec![MemoryLabel::Decision],
-                importance: 0.9,
-                pinned: true,
-                source: MemorySource::ThreadDistill,
-                scope: MemoryScope::Thread,
-                session_id: Some("session-123".to_string()),
-                thread_id: Some("thread-123".to_string()),
-                source_span: Some(MemorySourceSpan {
-                    start_turn_index: 2,
-                    end_turn_index: 4,
-                }),
-            })
-            .await
-            .expect("insert memory");
-
-        let reloaded = MemoryStore::new_with_record_path(backend, vdb, record_path);
-        let hits = reloaded
-            .search("memory retrieval", 5)
-            .await
-            .expect("search memories");
-        assert_eq!(hits[0].record.id, saved.id);
-        assert_eq!(hits[0].record.title, "Thread decision");
-        assert_eq!(hits[0].record.labels, vec![MemoryLabel::Decision]);
-        assert_eq!(hits[0].record.importance, 0.9);
-        assert!(hits[0].record.pinned);
-        assert_eq!(hits[0].record.source, MemorySource::ThreadDistill);
-        assert_eq!(hits[0].record.scope, MemoryScope::Thread);
-        assert_eq!(hits[0].record.session_id.as_deref(), Some("session-123"));
-        assert_eq!(hits[0].record.thread_id.as_deref(), Some("thread-123"));
-        assert_eq!(
-            hits[0].record.source_span,
-            Some(MemorySourceSpan {
-                start_turn_index: 2,
-                end_turn_index: 4,
-            })
-        );
-    }
-
-    #[test]
-    fn memory_record_retention_protects_pinned_user_created_and_high_importance_records() {
-        let mut record = test_memory_record("memory-retention-1", "Durable user note.");
-        assert!(record.is_protected_from_automatic_cleanup());
-
-        record.source = MemorySource::AgentTurn;
-        assert!(!record.is_protected_from_automatic_cleanup());
-
-        record.importance = HIGH_IMPORTANCE_RETENTION_THRESHOLD;
-        assert!(record.is_protected_from_automatic_cleanup());
-
-        record.importance = DEFAULT_IMPORTANCE;
-        record.pinned = true;
-        assert!(record.is_protected_from_automatic_cleanup());
-    }
-
-    #[tokio::test]
-    async fn memory_store_set_pinned_updates_and_persists_record_metadata() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let db_path = temp.path().join("lancedb");
-        let record_path = temp.path().join("memories").join("records.json");
-        let backend = Arc::new(MockLlm);
-        let vdb = Arc::new(VectorDB::new(db_path.to_str().expect("utf8 path")));
-        let store =
-            MemoryStore::new_with_record_path(backend.clone(), vdb.clone(), record_path.clone());
-
-        let saved = store
-            .insert(NewMemoryRecord::experience(
-                "Pinned records survive automatic cleanup.",
-            ))
-            .await
-            .expect("insert memory");
-        assert!(!saved.pinned);
-        assert!(!saved.is_protected_from_automatic_cleanup());
-
-        let pinned = store.set_pinned(&saved.id, true).await.expect("pin memory");
-        assert!(pinned.pinned);
-        assert!(pinned.is_protected_from_automatic_cleanup());
-
-        let reloaded = MemoryStore::new_with_record_path(backend, vdb, record_path);
-        let persisted = reloaded
-            .get(&saved.id)
-            .await
-            .expect("get memory")
-            .expect("persisted memory");
-        assert!(persisted.pinned);
-        assert!(persisted.is_protected_from_automatic_cleanup());
-    }
-
-    #[tokio::test]
-    async fn memory_store_updates_record_and_refreshes_search_index() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let db_path = temp.path().join("lancedb");
-        let record_path = temp.path().join("memories").join("records.json");
-        let backend = Arc::new(MockLlm);
-        let vdb = Arc::new(VectorDB::new(db_path.to_str().expect("utf8 path")));
-        let store =
-            MemoryStore::new_with_record_path(backend.clone(), vdb.clone(), record_path.clone());
-
-        let saved = store
-            .insert(NewMemoryRecord::experience("Old parser note."))
-            .await
-            .expect("insert memory");
-        let updated = store
-            .update(
-                &saved.id,
-                MemoryRecordPatch {
-                    title: Some("DSML parser decision".to_string()),
-                    content: Some(
-                        "DeepSeek DSML parser should use a structured parser.".to_string(),
-                    ),
-                    labels: Some(vec![MemoryLabel::Decision, MemoryLabel::Fact]),
-                    importance: Some(0.95),
-                    pinned: Some(true),
-                    scope: Some(MemoryScope::Workspace),
-                    ..Default::default()
-                },
-            )
-            .await
-            .expect("update memory");
-
-        assert_eq!(updated.id, saved.id);
-        assert_eq!(updated.title, "DSML parser decision");
-        assert_eq!(
-            updated.content,
-            "DeepSeek DSML parser should use a structured parser."
-        );
-        assert_eq!(
-            updated.labels,
-            vec![MemoryLabel::Decision, MemoryLabel::Fact]
-        );
-        assert_eq!(updated.importance, 0.95);
-        assert!(updated.pinned);
-        assert_eq!(updated.scope, MemoryScope::Workspace);
-
-        let reloaded = MemoryStore::new_with_record_path(backend, vdb, record_path);
-        let hits = reloaded
-            .search("structured parser", 5)
-            .await
-            .expect("search");
-        assert!(hits.iter().any(|hit| hit.record.id == saved.id));
-    }
-
-    #[tokio::test]
-    async fn memory_store_refreshes_index_metadata_for_scope_only_update() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let backend = Arc::new(MockLlm);
-        let vdb = Arc::new(VectorDB::new(temp.path().to_str().expect("utf8 path")));
-        let store = MemoryStore::new(backend.clone(), vdb.clone());
-
-        let saved = store
-            .insert(NewMemoryRecord::experience(
-                "Scope-only updates must refresh search index metadata.",
-            ))
-            .await
-            .expect("insert memory");
-        store
-            .update(
-                &saved.id,
-                MemoryRecordPatch {
-                    scope: Some(MemoryScope::Workspace),
-                    ..Default::default()
-                },
-            )
-            .await
-            .expect("update scope");
-
-        let metadata_hits = vdb
-            .search_with_metadata(
-                EXPERIENCES_TABLE,
-                LlmBackend::embed(backend.as_ref(), "Scope-only updates")
-                    .await
-                    .expect("embed"),
-                5,
-            )
-            .await
-            .expect("search metadata");
-        let metadata = metadata_hits
-            .into_iter()
-            .map(|(metadata, _score)| metadata)
-            .find(|metadata| metadata.id.as_deref() == Some(saved.id.as_str()))
-            .expect("updated index metadata");
-        assert_eq!(
-            metadata.session_id,
-            memory_scope_key(&MemoryScope::Workspace)
-        );
-    }
-
-    #[tokio::test]
-    async fn memory_store_delete_hides_stale_lancedb_hits() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let backend = Arc::new(MockLlm);
-        let vdb = Arc::new(VectorDB::new(temp.path().to_str().expect("utf8 path")));
-        let store = MemoryStore::new(backend, vdb);
-
-        let saved = store
-            .insert(NewMemoryRecord::experience(
-                "Deleted memory should not be rehydrated from stale LanceDB rows.",
-            ))
-            .await
-            .expect("insert memory");
-        assert_eq!(
-            store.delete(&saved.id).await.expect("delete memory"),
-            Some(saved.clone())
-        );
-        assert_eq!(store.get(&saved.id).await.expect("get deleted"), None);
-
-        let hits = store.search("stale LanceDB rows", 5).await.expect("search");
-        assert!(
-            hits.iter().all(|hit| hit.record.id != saved.id),
-            "deleted domain records must not be reconstructed from stale LanceDB index rows"
-        );
-    }
-
-    #[tokio::test]
-    async fn memory_store_list_labels_counts_records_by_scope() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let store = MemoryStore::new(
-            Arc::new(MockLlm),
-            Arc::new(VectorDB::new(temp.path().to_str().expect("utf8 path"))),
-        );
-
-        store
-            .insert(NewMemoryRecord {
-                title: Some("Workspace decision".to_string()),
-                content: "Use the shared MemoryStore API.".to_string(),
-                labels: vec![MemoryLabel::Decision, MemoryLabel::Fact],
-                importance: 0.7,
-                pinned: false,
-                source: MemorySource::UserCreated,
-                scope: MemoryScope::Workspace,
-                session_id: None,
-                thread_id: None,
-                source_span: None,
-            })
-            .await
-            .expect("insert workspace memory");
-        store
-            .insert(NewMemoryRecord {
-                title: Some("Thread fact".to_string()),
-                content: "Thread facts stay scoped to the thread.".to_string(),
-                labels: vec![MemoryLabel::Fact],
-                importance: 0.4,
-                pinned: false,
-                source: MemorySource::ThreadDistill,
-                scope: MemoryScope::Thread,
-                session_id: Some("session-1".to_string()),
-                thread_id: Some("thread-1".to_string()),
-                source_span: None,
-            })
-            .await
-            .expect("insert thread memory");
-
-        assert_eq!(
-            store.list_labels(None).await.expect("all labels"),
-            vec![
-                MemoryLabelCount {
-                    label: MemoryLabel::Fact,
-                    count: 2,
-                },
-                MemoryLabelCount {
-                    label: MemoryLabel::Decision,
-                    count: 1,
-                },
-            ]
-        );
-        assert_eq!(
-            store
-                .list_labels(Some(MemoryScope::Workspace))
-                .await
-                .expect("workspace labels"),
-            vec![
-                MemoryLabelCount {
-                    label: MemoryLabel::Decision,
-                    count: 1,
-                },
-                MemoryLabelCount {
-                    label: MemoryLabel::Fact,
-                    count: 1,
-                },
-            ]
-        );
-    }
+    include!("memory_store_tests.rs");
 }
